btc-paper-trader/pyproject.toml
[project]
name = "btc-paper-trader"
version = "0.1.0"
description = "BTC/USD paper trading system for autotrader model validation"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.0.0",
    "pandas>=2.2.0",
    "scikit-learn>=1.5.0",
    "pyarrow>=15.0.0",
    "requests>=2.32.0",
    "websockets>=12.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["src*"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
